Remove need for `SharedPartnerAuthViewModel`
